refactor: migrate primary keys to long and enhance Result pattern
  - Change database primary key type from ulong (numeric(20,0)) to long (bigint)
  - Update entity models, DTOs, and migrations to use long for IDs
  - Add Match extension method to Result pattern for functional error handling
  - Refactor ClaimsPrincipal mapping to use extension methods instead of Mapster
  - Add UseIdentityColumn() to entities configuration
  - Initialize Guid property with default value in EntityBase
  - Comment out incomplete role assignment in TokenExchangeService
  - Disable v0.1 data seeding migration temporarily
  - Remove previous migrations to keep the sample simple and understandable

+++ Source/Data/Data/Db/Sql/Common/Seed/Seed_v1.0.sql
@@ -4,7 +4,7 @@ DO
 $$
 DECLARE
     currencies      TEXT[] := ARRAY ['USD', 'SGD', 'MYR'];
-    user_ids        NUMERIC(20,0)[];
+    user_ids        BIGINT[];
     i               BIGINT;
     random_text     TEXT;
     category_val    INT;
